Initial Game without database included.
Both the data stream and game must be run at the same time. The data stream loads movement commands into the database from the Epoc, and the Owl reads the corresponding movements and reacts to them. These movements are based on your eye movements, since there was too much noise and training required to have it done with my thoughts alone.